Backend/Team-Software-Project-2025-YanHaoSun/Recommendation/src/main/java/org/recommendation/Rec_LLM_Module/requirements.txt
# 基础科学计算
numpy>=1.24.0
pandas>=2.0.0 

# 环境配置
python-dotenv>=1.0.0

# OpenAI API
openai>=1.0.0

# FastAPI 核心框架
fastapi>=0.104.0
uvicorn[standard]>=0.24.0

# 数据验证和序列化
pydantic>=2.0.0

# 数据库相关
sqlalchemy>=2.0.0
aiosqlite>=0.19.0

# Java集成数据库驱动
pymysql>=1.1.0  # MySQL驱动
psycopg2-binary>=2.9.0  # PostgreSQL驱动

# HTTP 客户端（用于调用Java后端API）
httpx>=0.25.0
requests>=2.31.0

# 日期时间处理
python-dateutil>=2.8.2

# 日志处理
loguru>=0.7.0

# 测试相关（开发时使用）
pytest>=7.4.0
pytest-asyncio>=0.21.0

# 类型检查（开发时使用）
mypy>=1.5.0

# 数据处理和分析
scikit-learn>=1.3.0

# 文件处理
aiofiles>=23.0.0